// src/types/user.type.tsx
type UserStore = {
    id: number
    email: string
    firstName?: string
    lastName?: string
    username?: string
    gender?: string
    image?: string
    role?: string
    accessToken?: string
    refreshToken?: string
    isCheckingUser?: boolean
    clearUser: () => void
    setUser: (user: Partial<UserStore>) => void
}


type User = {
    id: number
    firstName: string
    lastName: string
    maidenName: string
    age: number
    gender: string
    email: string
    phone: string
    username: string
    password: string
    birthDate: string
    image: string
    bloodGroup: string
    height: number
    weight: number
    eyeColor: string
    hair: { color: string; type: string }
    ip: string
    address: {
        address: string
        city: string
        state: string
        stateCode: string
        postalCode?: string
        coordinates?: { lat: number; lng: number }
    }
    macAddress: string
    university: string
    bank: {
        cardExpire: string
        cardNumber: string
        cardType: string
        currency: string
        iban: string
    }
    company: {
        department: string
        name: string
        title: string
        address?: any
    }
    ein: string
    ssn: string
    userAgent: string
    crypto: {
        coin: string
        wallet: string
        network: string
    }
    role: string
}

export type { UserStore, User }